requirements.txt
crewai==0.5.0
crewai-tools==0.0.8
langchain==0.1.13
langchain-openai==0.1.1
langchain-community==0.0.29
pypdf==4.1.0
PyMuPDF==1.23.24
sentence-transformers==2.3.1
faiss-cpu==1.7.4
chromadb==0.4.24
click==8.1.7
python-dotenv==1.0.1
pydantic==2.6.4
tiktoken==0.6.0
numpy==1.24.3
pandas==2.0.1